+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - `KML` as a built in media type ([#1127](https://github.com/stac-utils/pystac/issues/1127))
 - `move/copy/delete` operations for local Assets ([#1158](https://github.com/stac-utils/pystac/issues/1158))
 - Latest core STAC spec jsonshemas are included in pytstac and used for validation ([#1165](https://github.com/stac-utils/pystac/pull/1165))
+- Xarray Assets Extension class ([#1161](https://github.com/stac-utils/pystac/pull/1161))
 
 ### Changed